+++ allauth/socialaccount/internal/flows/connect.py
@@ -1,14 +1,14 @@
 from django.contrib import messages
+from django.core.exceptions import PermissionDenied
 from django.http import HttpResponseRedirect
 
 from allauth import app_settings as allauth_settings
 from allauth.account import app_settings as account_settings
 from allauth.account.adapter import get_adapter as get_account_adapter
+from allauth.account.internal import flows
 from allauth.account.models import EmailAddress
-from allauth.account.reauthentication import (
-    raise_if_reauthentication_required,
-    reauthenticate_then_callback,
-)
+from allauth.account.reauthentication import raise_if_reauthentication_required
+from allauth.core.exceptions import ReauthenticationRequired
 from allauth.socialaccount import signals
 from allauth.socialaccount.adapter import get_adapter
 from allauth.socialaccount.models import SocialAccount, SocialLogin
@@ -70,45 +70,22 @@ def resume_connect(request, serialized_state):
 
 
 def connect(request, sociallogin):
-    if request.user.is_anonymous:
+    try:
+        ok, action, message = do_connect(request, sociallogin)
+    except PermissionDenied:
         # This should not happen. Simply redirect to the connections
         # view (which has a login required)
         connect_redirect_url = get_adapter().get_connect_redirect_url(
             request, sociallogin.account
         )
         return HttpResponseRedirect(connect_redirect_url)
-    if account_settings.REAUTHENTICATION_REQUIRED:
-        response = reauthenticate_then_callback(
+    except ReauthenticationRequired:
+        return flows.reauthentication.stash_and_reauthenticate(
             request,
-            lambda request: sociallogin.serialize(),
+            sociallogin.serialize(request),
             "allauth.socialaccount.internal.flows.connect.resume_connect",
         )
-        if response:
-            return response
-    level = messages.INFO
-    message = "socialaccount/messages/account_connected.txt"
-    action = None
-    if sociallogin.is_existing:
-        if sociallogin.user != request.user:
-            # Social account of other user. For now, this scenario
-            # is not supported. Issue is that one cannot simply
-            # remove the social account from the other user, as
-            # that may render the account unusable.
-            level = messages.ERROR
-            message = "socialaccount/messages/account_connected_other.txt"
-        elif not sociallogin.account._state.adding:
-            # This account is already connected -- we give the opportunity
-            # for customized behaviour through use of a signal.
-            action = "updated"
-            message = "socialaccount/messages/account_connected_updated.txt"
-        else:
-            action = "added"
-            sociallogin.connect(request, request.user)
-    else:
-        # New account, let's connect
-        action = "added"
-        sociallogin.connect(request, request.user)
-    assert request.user.is_authenticated
+    level = messages.INFO if ok else messages.ERROR
     default_next = get_adapter().get_connect_redirect_url(request, sociallogin.account)
     next_url = sociallogin.get_redirect_url(request) or default_next
     get_account_adapter(request).add_message(
@@ -118,3 +95,32 @@ def connect(request, sociallogin):
         message_context={"sociallogin": sociallogin, "action": action},
     )
     return HttpResponseRedirect(next_url)
+
+
+def do_connect(request, sociallogin):
+    if request.user.is_anonymous:
+        raise PermissionDenied()
+    if account_settings.REAUTHENTICATION_REQUIRED:
+        raise_if_reauthentication_required(request)
+    message = "socialaccount/messages/account_connected.txt"
+    action = None
+    ok = True
+    if sociallogin.is_existing:
+        if sociallogin.user != request.user:
+            # Social account of other user. For now, this scenario
+            # is not supported. Issue is that one cannot simply
+            # remove the social account from the other user, as
+            # that may render the account unusable.
+            message = "socialaccount/messages/account_connected_other.txt"
+            ok = False
+        elif not sociallogin.account._state.adding:
+            action = "updated"
+            message = "socialaccount/messages/account_connected_updated.txt"
+        else:
+            action = "added"
+            sociallogin.connect(request, request.user)
+    else:
+        # New account, let's connect
+        action = "added"
+        sociallogin.connect(request, request.user)
+    return ok, action, message