+++ ChangeLog.rst
@@ -4,6 +4,9 @@
 - ``allauth.headless`` now supports the ``is_open_for_signup()`` adapter method.
   In case signup is closed, a 403 is returned during signup.
 
+- Connecting a third-party account in ``HEADLESS_ONLY`` mode failed if the
+  connections view could not be reversed, fixed.
+
 
 0.63.1 (2024-05-17)
 *******************